+++ besser_backend/models/__init__.py
@@ -1,0 +1,3 @@
+from .class_diagram import ClassDiagramInput
+
+__all__ = ['ClassDiagramInput']
