+++ besser_backend/models/class_diagram.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class ClassDiagramInput(BaseModel):
+    elements: dict
+    generator: str
