+++ besser_backend/utils/__init__.py
@@ -1,0 +1,13 @@
+from .constants import (
+    VALID_PRIMITIVE_TYPES,
+    VISIBILITY_MAP,
+    RELATIONSHIP_TYPES
+)
+
+
+__all__ = [
+    # From constants
+    'VALID_PRIMITIVE_TYPES',
+    'VISIBILITY_MAP',
+    'RELATIONSHIP_TYPES',
+]
