+++ besser_backend/utils/constants.py
@@ -1,0 +1,26 @@
+VALID_PRIMITIVE_TYPES = {
+    "str": "str",
+    "string": "str",
+    "int": "int",
+    "integer": "int",
+    "float": "float",
+    "double": "float",
+    "bool": "bool",
+    "boolean": "bool",
+    "date": "date",
+    "datetime": "datetime"
+}
+
+VISIBILITY_MAP = {
+    "+": "public",
+    "-": "private",
+    "#": "protected",
+    "~": "package"
+}
+
+RELATIONSHIP_TYPES = {
+    "bidirectional": "ClassBidirectional",
+    "unidirectional": "ClassUnidirectional",
+    "composition": "ClassComposition",
+    "inheritance": "ClassInheritance"
+}
